move maps to separate files
