+++ backuper/core.py
@@ -13,7 +13,7 @@ from typing import Any, TypeVar
 from pydantic import BaseModel
 
 from backuper import config
-from backuper.models import backup_target_models, upload_provider_models
+from backuper.models import backup_target_models, models_mapping, upload_provider_models
 
 log = logging.getLogger(__name__)
 
@@ -48,25 +48,6 @@ def run_subprocess(shell_args: str) -> str:
     log.debug("run_subprocess stdout: %s", p.stdout)
     log.debug("run_subprocess stderr: %s", p.stderr)
     return p.stdout
-
-
-def get_target_map() -> dict[str, type[backup_target_models.TargetModel]]:
-    return {
-        config.BackupTargetEnum.FILE: backup_target_models.SingleFileTargetModel,
-        config.BackupTargetEnum.FOLDER: backup_target_models.DirectoryTargetModel,
-        config.BackupTargetEnum.MARIADB: backup_target_models.MariaDBTargetModel,
-        config.BackupTargetEnum.MYSQL: backup_target_models.MySQLTargetModel,
-        config.BackupTargetEnum.POSTGRESQL: backup_target_models.PostgreSQLTargetModel,
-    }
-
-
-def get_provider_map() -> dict[str, type[upload_provider_models.ProviderModel]]:
-    return {
-        config.UploadProviderEnum.AZURE: upload_provider_models.AzureProviderModel,
-        config.UploadProviderEnum.LOCAL_FILES_DEBUG: upload_provider_models.DebugProviderModel,
-        config.UploadProviderEnum.GOOGLE_CLOUD_STORAGE: upload_provider_models.GCSProviderModel,
-        config.UploadProviderEnum.AWS_S3: upload_provider_models.AWSProviderModel,
-    }
 
 
 def remove_path(path: Path) -> None:
@@ -164,7 +145,7 @@ def _validate_model(
 
 
 def create_target_models() -> list[backup_target_models.TargetModel]:
-    target_map = get_target_map()
+    target_map = models_mapping.get_target_map()
 
     targets: list[backup_target_models.TargetModel] = []
     for env_name, env_value in os.environ.items():
@@ -182,7 +163,7 @@ def create_target_models() -> list[backup_target_models.TargetModel]:
 
 
 def create_provider_model() -> upload_provider_models.ProviderModel:
-    provider_map = get_provider_map()
+    provider_map = models_mapping.get_provider_map()
 
     log.info("start validating BACKUP_PROVIDER environment variable")
